Improve getting started instructions
This patch improves the getting started instructions by adding a
section on how to install the dependencies and migrate the database.

The `git` commands are also updated to stop making use of the `docked`
alias as it seems to cause issues on some systems.

The alias is also recommended to be added to the user's shell
configuration file.