Introduce Stimulus JS to double click on a todo and edit it
This patch introduces a new `visitable` Stimulus JS controller to handle
the `dblclick` event on a todo and visit the `todos#edit` routes which
will render the `edit.turbo_stream.erb` template and replace the todo
with the form to edit it.

App specification
-----------------

- https://github.com/tastejs/todomvc/blob/master/app-spec.md#item
- https://github.com/tastejs/todomvc/blob/master/app-spec.md#editing

Documentation
-------------

- https://hotwired.dev
- https://stimulus.hotwired.dev/handbook/introduction
- https://stimulus.hotwired.dev/reference/actions
- https://developer.mozilla.org/en-US/docs/Web/API/Element/dblclick_event
- https://stimulus.hotwired.dev/reference/values
- https://turbo.hotwired.dev/reference/drive
- https://turbo.hotwired.dev/reference/streams

Run the tests
-------------

The exercise's test is located in the `test/system/todos_test.rb` file.

```bash
docked rails test test/system/todos_test.rb:83
```