Submit editing form when focusing out (blur)
This patch ensures that the editing form is submitted when the focus on
the input is lost (`blur()` event).

A new `SubmitableController` generic Stimulus controller is created to
handle the `blur` event on the input. The `requestSubmit` action is
called on the form element when the `blur` event is triggered.

App specification
-----------------

- https://github.com/tastejs/todomvc/blob/master/app-spec.md#editing

Documentation
-------------

- https://stimulus.hotwired.dev/reference/actions
- https://developer.mozilla.org/en-US/docs/Web/API/Window/blur_event
- https://developer.mozilla.org/en-US/docs/Web/API/HTMLFormElement/requestSubmit

Run the tests
-------------

The exercise's test is located in the `test/system/todos_test.rb` file.

```bash
docked rails test test/system/todos_test.rb:97
```